requirements.txt
streamlit==1.37.1
pandas==2.2.1
matplotlib==3.8.3
seaborn==0.13.2
Pillow==9.0.1
